- Removed the support for Complex numbers.
If anyone is interested in adding it back, the file is avaiable inside the "experiments" directory. However, I strongly advise against it. The main reason was because MPC uses MPFR for the internal storage, while BigNum uses GMPq, which may have leaded to many conversions problems in the long run. Therefore, it's not part of this distribution anymore.

- Improved the Inf type.
- Extended (b)lsft and (b)rsft methods to support negative integers.
- Minor documentation fixes and improvents.
- Added more tests.